Added unit test.
